prototype of decoding a simple key/value payload.
Client's send key/value pairs where each key and value is prefixed with
a 2-byte (16-bit) length (in network byte order) followed by the key or
value. It's assumed the key and value are opaque and not C-strings, so
no NUL byte terminator is expected as a seperator or suffix.